requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.5.0
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
PyAudio==0.2.14
pydantic==2.9.2
pydantic_core==2.23.4
regex==2024.9.11
requests==2.32.3
semantic-version==2.10.0
setuptools==75.1.0
setuptools-rust==1.10.1
six==1.16.0
sniffio==1.3.1
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
